refactor: 모델 앙상블을 위한 기존모델 및 api 구조변경
